handle tri-meshes and meshes > 100000 verts
Tri meshes must be saved different not to fool blender (other wise
subdivision surface will work incorrect and also selection with ctrl-+
will freeze system (eventually).

indices > 100000 were saved as 1e05 before. save-json now works
different for numbers > 100000 (because of numpy, the vertex index is used
internally as a float and will create trouble in blender when it is
decoded as 1e05).